src/test/java/com/br/joaoptgaino/loans/controller/LoanControllerTest.java
package com.br.joaoptgaino.loans.controller;

import com.br.joaoptgaino.loans.dto.loans.LoanDTO;
import com.br.joaoptgaino.loans.model.enums.LoanType;
import com.br.joaoptgaino.loans.service.LoanService;
import com.google.gson.Gson;
import com.google.gson.GsonBuilder;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.springframework.data.web.PageableHandlerMethodArgumentResolver;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.RequestBuilder;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.List;

import static com.br.joaoptgaino.loans.fixtures.LoanFixture.getLoanDTO;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@ExtendWith(SpringExtension.class)
public class LoanControllerTest {
    private MockMvc mockMvc;
    private final Gson gson = new GsonBuilder().create();

    private static final String BASE_URL = "/loans";
    @Mock
    private LoanService loanService;

    @BeforeEach
    void setup() {
        mockMvc = MockMvcBuilders
                .standaloneSetup(new LoanController(loanService))
                .setCustomArgumentResolvers(new PageableHandlerMethodArgumentResolver())
                .build();
    }

    @Test
    public void findAllLoans() throws Exception {
        LoanDTO loanDTO = getLoanDTO(LoanType.PERSONAL);
        when(loanService.findAll()).thenReturn(List.of(loanDTO));

        RequestBuilder request = MockMvcRequestBuilders.get(BASE_URL)
                .accept(MediaType.APPLICATION_JSON);
        mockMvc.perform(request)
                .andExpect(status().isOk())
                .andExpect(MockMvcResultMatchers.jsonPath("$").exists())
                .andExpect(MockMvcResultMatchers.jsonPath("$").isArray());
    }
}
